+++ sei-ibc-go/modules/core/02-client/client/cli/cli.go
@@ -23,7 +23,7 @@ func GetQueryCmd() *cobra.Command {
 		GetCmdQueryConsensusStates(),
 		GetCmdQueryConsensusState(),
 		GetCmdQueryHeader(),
-		GetCmdNodeConsensusState(),
+		GetCmdSelfConsensusState(),
 		GetCmdParams(),
 	)